object_detector/requirements.txt
torch
torchvision
numpy
matplotlib
Pillow
pycocotools
opencv-python
scikit-learn
